Reimplement pkg/evaluation to support parallel evaluation
Up until this point bonanza_builder would run fully sequentially.
Furthermore, evaluation of analysis steps would always be interleaved
with uploading of Merkle tree outputs into the CAS. This is obviously
not a good strategy.

This change rewrites FullyComputeValue entirely to allow parallel
evaluation. We can now launch a pool of goroutines to do the
computation. Furthermore, it eliminates storage related handling as part
of the evaluation code. We now simply pass in an ObjectManager to
convert from references to metadata and vice versa.

A downside of this change is that memory usage of bonanza_builder is now
a lot higher. This is because our ObjectManager implementation is still
far too trivial, in that it does not proactively flush any Merkle tree
nodes to storage. Also, once flushing has been performed, the object
contents still don't get evicted. The next step will be to provide a
more definitive/complete implementation.

While there, we change the evaluation code to keep track of SHA-256 sums
of keys, as opposed to storing full keys. In addition to reducing memory
usage significantly, it reduces the amount of time spent holding locks,
as we can now do map lookups on [32]byte.