Permit rule(cfg = "target")
I was under the impression that rule(cfg = ...) could only take user
defined transitions. Well, it turns out that the rule builder used by
rules_python uses "target" as a default value, as opposed to None.
Considering that cfg = "target" is already intended as a no-op, make it
the internal default value. Change EncodeUserDefinedTransition() to
return a nil message.