Make the difference between (un)configured target references explicit
Right now TargetReference is in a bit of an odd shape, where depending
on whether it's configured the .label field refers to the originally
provided label, or the one after aliases have been expanded.

Let's add an .original_label field that's always available, even if a
target is not configured. Only when it is available do we offer .label
and access to the provider instances.