Switch to AES-GCM-SIV for encrypting objects
This means we now use authenticating encryption. This likely doesn't
have a huge value over what we already have,considering that a Merkle
tree is already authentication. However, because we place a hash of the
binary encoder configurations in the additional data we can now let
bonanza_browser automatically pick the right decoder.