Add inlinedtree.Capturing()
There are many cases in which we want to use inlinedtree.Build() in
combination with model_core.CreatedObjectCapturer. Right now this
requires us to make explicit calls into the capturer from within the
ParentAppender.

Let's add a utility function that does this for us before calling into
the ParentAppender. That way the implementation of those functions can
remain simple, even when we add support for letting
CreatedObjectCapturer fail.