Get rid of CloneableReferenceMetadata
All of this hinges on the fact that we can make ReferenceMetadata
implementations cloneable. Giving it more thought, I don't think this is
desirable or realistic. By not allowing them to be cloneable, we can
also work towards doing proper reference count checking for them and
bailing out of we leak any metadata.