Let CapturedParentNodeComputer take a Message[T, LocalReference]
There's one spot where we don't use CapturedParentNodeComputer, and
that's because we need to get access to the outgoing references of the
object. Only a plain ParentNodeComputer provides that. However, we can
cleanly extend CapturedParentNodeComputer to offer this by turning this
into a Message[T, LocalReference]. This also tends to make the
implementations of these functions nicer, as they can simply use
model_core.Nested() to extract nested fields.