Add metadata leak checking to bonanza_builder
Right now there are various places in which we leak ReferenceMetadata
objects. This is suboptimal, as they may be backed by resources that
need to be released explicitly.

Let's add a decorator to our implementation of Computer that validates
that ReferenceMetadata objects don't leak. This currently causes almost
immediate build failures. Let's try to get those fixed soon.