Add repository_ctx.workspace_root
It looks like this is used by aspect_tools_telemetry. Adding this
requires us to reshuffle some code, because it means we can no longer
let repository_ctx be an ordinary struct. It now needs to be a custom
type, as repository_ctx.workspace_root needs to perform some work, while
it is not a method.