Get rid of btree.SplitBuilder
I would like to investigate whether we can make the btree code aware of
heights. It would be nice if higher elements can also be placed higher
in the tree. This would result the overall height of the resulting tree.

This change gets rid of the minimumCount parameters of ProllyChunker and
replaces it with an isParent() function. This can be used to either
choose between 1 or 2 as a lowerbound on the number of elements.