Move responsibility of merging nodes to UniformBuilder
I'd like to investigate whether I can add an implementation that is more
like a true B-tree instead of a B+ tree, where leaves may also get
stored further up in the tree. This in an attempt to reduce the height
of DAGs. However, this will require us to extract the nodes from a
Chunker without having them be merged in the process.