Make B-tree generation aware of the height of nodes
In our case we often need to create B-trees of B-trees of B-trees of
[...]. Even though each of the individual B-trees is uniform in height,
the resulting object graph is not. This can lead to some unreasonably
high graphs. For example, the full evaluation graph for a build of
//cmd/bb_storage leads to a graph of height 41.

This change brings in a new algorithm for constructing B-trees, which
takes the height of the nodes it stores into account. If a node has a
higher height, it will also be stored higher up in the B-tree. This
means that B-tree objects may now contain a mixture of parent and leaf
nodes, as opposed to containing only the former or the latter.

When implemented naively, this approach may lead to B-trees that contain
objects that are smaller than the desired minimum size. This is
something we absolutely want to avoid, which is why we add the new
threshold 'PopLargeEnough', which prevents exactly that. Any nodes that
cannot be stored because they reside in an object that is too small are
combined with objects from higher levels.

Benchmarks have shown that this change is capable of reducing the height
of the aforementioned evaluation graph from 41 levels to 21. As objects
still respect the desired minimum size, the overall object count should
only increase slightly.