Make use_repo_rule() work if the identifier is an alias
When invoking actual module extensions, it's guaranteed that the
repository rule identifier that we store is the canonical name of the
repository rule. For repository rules that are invoked directly from
MODULE.bazel using use_repo_rule(), this is not the case.

We can solve this by calling RepositoryRuleObject recursively. This
cannot not lead to infinite recursion, as the evaluation framework will
detect cyclic dependencies.