After printing child entries, re-enable white prefix rendering because the prefix changed.
In a listing like:

- .git/FETCH_HEAD
- .git/HEAD
- .git/hooks/update.sample
- .git/index

For HEAD, prefix path would be printed as gray since it's the same as the
previous entry, FETCH_HEAD. Earlier, 'index' would also be printed as gray
because it's not the first file in this prefix. But it should be printed as
white, since it is in a different directory from the listing before it.
Therefore, reset firstPrefix to true when we've had any entries from
subdirectories.

Currently, len(children) is always >0 but this may not be the case in the
future, so check it anyway.