Allow FileFactory.LookupFile() to fail
I'd like to be able to use ObjectBackedInitialContentsFetcher to create
input roots that are fully writable. This means that for every file that
is being looked up, a file needs to be drawn from the FilePool. As
FilePool.NewFile() may fail, we need to be able to propagate such
errors.