Make files in input roots of repo rules mutable
For regular build actions we want all input files to be read-only. This
is the easiest to implement. It is also how Bazel and Buildbarn behave.
However, for repository rules we need input files to be writable. There
are many steps that replace or append data to existing files.

This change adds a new flag to command actions that clients can use to
request writable input files. When set, a different implementation of
FileFactory is used that, instead of regular object backed files,
creates files backed by FilePool.

Fixes: #4