Change the encoding of FileContents to support storing holes
This commit changes the definition of FileContents as follows:

- In addition to nested lists and chunks, FileContents can now declare
  regions of a file as being a hole. As there is no data associated with
  a hole, we just store a google.protobuf.Empty. Holes can't have other
  holes as successors/predecessors.

- We explicitly state that actual chunks of data in a file MUST have at
  least one non-zero byte in them.

- We turn file_contents_list_reference into a proper message instead of
  just a reference. This allows us to store additional properties. We
  immediately make use of this by storing whether a file contents list
  is sparse (i.e., containing one or more holes).

With these changes in place, we are not actually capable of literally
preserving the holes in a file. However, any sufficiently large region
that consists only of null bytes is automatically turned into a hole. We
rely on the content defined chunking algorithm to determine which ranges
are suitable for turning into holes.

This change requires us to rework FileContentsEntry. I've added a
contiguousCount field, which will keep track of the next index at which
a hole is present. This makes it possible to efficiently implement
lseek(SEEK_HOLE) without doing a full scan. A negative value is used to
indicate that the current entry corresponds to a hole, and thus the
reference field is not valid. Because this makes dealing with
FileContentsEntry more error prone, I've used the occasion to make all
members private and add more helper functions for constructing and
accessing this type.

The next step will be to actually patch up implementations of
VirtualSeek() and GetNextRegionOffset() to make use of this information.
It would also be nice if we could somehow optimize
CreateFileMerkleTree() to be aware of sparseness in the input file as
well.