Overhaul the design of the Tag Store
When I initially designed the Tag Store, I didn't give any consideration
to privacy and security. Now that we've addressed this specific topic
for the Object Store, it's time to give the Tag Store the same
treatment.

In this commit we change keys of tags from google.protobuf.Any to a
tuple of an elliptic-curve public key and a use case specific hash. The
elliptic-curve key is used to sign the values of tags. This means that
any system is capable of validating signatures. It's just that to create
new tags you need to have the private key. This is great, because it
means we now have yet another case where we change an authorization
problem to a key distribution problem.