Add a local storage backend for tags based a lossy hash map
This change adds a generic configuration message for our lossy hash map.
That way local object storage, the leases map, and tag storage are all
configured through the same means.

While there, rework the tag key and value messages to use static arrays,
as opposed to slices. The keys and signatures are small enough that it
makes sense to just put everything in a single message.