Remove custom CompareSameType() function for attr.*()
In the past I stumbled upon some Bazel rules that depended on literal
comparisons of attr.*() objects. So something like this:

    if attr.string(values = [...]) == attr.string(values = [...]):
        ...

To make this work I added a very crooked comparison function to
model_starlark.Attr. Because attribute objects may be backed by Starlark
definitions, this means that we potentially had to encode values first.

Now with the latest version of the 'bonanza' branch of bb-storage I am
unable to find code that actually exercises. This is good, because I
think that comparing attr.*() objects is bad. attr.label(cfg = ...) can
take a user-defined transition. And user-defined transitions can have
implementation functions that are anonymous functions. This means that
comparing these objects may actuall involve anonymous functions, which
should be avoided.

Fixes: #19