Rename BinaryEncoder to DeterministicBinaryEncoder
When using the tag store, it will not be possible to use an encryption
scheme that only considers the input data (and the key). The reason
being that the tag store will not be capable of storing decoding
parameters. Decoding parameters are not part of the storage layer at
all.

To solve that, I want to make it so that objects referenced by tags
actually make use of nonces. This means that EncodeBinary() will need to
have different signature for such objects. The shape of the encrypted
objects will also be different, meaning we're going to need a completely
different implementation of the encrypting binary encoder.