Add KeyedBinaryEncoder
The Tag Store won't allow us to store decoding parameters. On one hand
this is also good, as storing it in literal form it would allow someone
who can scan all tags in storage to immediately decrypt all results.
This is why for top-level objects referenced by tags we should use a
different scheme.

This change adds a new KeyedBinaryEncoder interface. The EncodeBinary()
function is identical to that of DeterministicBinaryEncoder, except that
it allows the desired decoding parameters to be passed in. In our case
we let this be the AES-GCM-SIV nonce.

While there, rename BinaryEncoderList to
EncryptingBinaryEncoderAdditionalDataInput, and let it contain an enum
value that indicates what kind of BinaryEncoder was being used to encode
an object. That way we can be certain that any mixups cause immediate
decryption/verification failures.