Add model_parser.MessageObjectReader[TReference, TMessage]
In a lot of places where we use ParsedObjectReader we're interested in
passing in decodable references, and access the object's contents using
our Message type. Let's add a type alias for this. In addition to making
the code easier to read, it may also become easier to refactor if we
ever need to make some fundamental API changes.