Port MutableProtoStore from bb-remote-execution
In our case we need to implement it on top of Object Store and Tag
Store. To improve security, we don't use the provided tag key hash
literally. Instead, we first compute a SHA-512 hash of it, so that we
can use the second half of that hash as the decoding parameters. This
ensures that someone with access to the storage nodes isn't capable of
walking all tags and decrypting the objects they reference if the
encryption key gets leaked.