Switch to a different implementation of AES-GCM-SIV
These two libraries seem to be compatible with each other in terms of
API and behavior. The advantage of the new one is that it also has an
implementation that is optimized for ARM64.

Hopefully we're not going to need this in the future, as there is a
reasonable chance the Go standard library is going to include support
for AES-GCM-SIV directly.