Reuse PatchedMessage for computing tag key hashes
So far I was planning on using SHA-256 hashes of messages and embedding
those into other messages. However, instead of doing that we could just
leverage our existing tools for creating messages with references to
other objects. This has the advantage that regular messages with
references can be embedded into tag keys.