Fix crashes when object refreshing is performed with persistency
Calling EpochList.GetCurrentEpochState() from within the read code path
is problematic, because we might not even have an epoch. You could also
argue that refreshing doesn't necessarily need to be based on the state
of epochs, but more on the offset tracked by LocationBlobMap.
LocationBlobMap's offset could be far ahead of that of the epoch lists.

Because LocationBlobMap is unaware of a 'minimum location', invert the
logic we have right now to calculate backwards.