Rework RecursiveComputer to support queueing for uploading results
This adds a new set of counters to the progress update message, as we
now want to report the number of keys that are (queued for) uploading as
well. Also rename some of the queuing data types to make it explicit
that these are for queueing evaluations, not queueing uploads.