Experiment with adding Graphlets
Placing all evaluation results in a single list has the downside that
every build needs to recompute a supermassive B-tree. This makes it
virtually impossible to offer good performance when most results come
from the cache. It also means that changes that are spread out across
the key space in a uniform manner cause a large part of the B-tree to
change.

My idea for solving this is to model the evaluation results not as a
flat namespace. Instead, I would like to recursively partition the build
graph using various amounts of precision. At the top level we have a
list of the coarsest parts of the graph. Parts of the graph will be
called Graphlets.