Make cache lookups succeed for keys without dependencies
Keys without dependencies are very rare, especially if they don't have
any overrides in place. One way to support them would be to let the
initial lookup return an empty value. However, that means we then need
to support empty sets of additional dependencies. Instead of doing that,
let's allow the initial lookup to return a value as well. This means we
also stick to one round trip in that case.